Improver orchestrator 2026 04 23 00 00 00

+++ include/functions.us.h
@@ -278,7 +278,7 @@ s32 func_8007D2B0_4D760(void);
 s32 func_80082CBC_5316C(s32 arg0, s32* arg1);
 s32 func_80076504_469B4(void);
 s32 func_800776BC_47B6C(void);
-extern MissionData* func_800706E8_40B98(s16 arg0);
+extern MissionData* func_800706E8_40B98(s32 arg0);
 void func_8007949C_4994C(s32 arg0, s32 arg1);
 void func_8007A754_4AC04(void);
 void func_8007A774_4AC24(s32 arg0);
